Tcga Presets (wip) - load from to db

+++ dataset/domain/src/main/java/edu/dfci/cccb/mev/dataset/domain/jooq/JooqBasedDataSourceValues.java
@@ -39,17 +39,20 @@ public class JooqBasedDataSourceValues extends AbstractDataSourceValues {
   private final Field<String> row;
   private final Field<String> column;
   private final Field<Double> value;
+  private final boolean isTemporary;
 
   public JooqBasedDataSourceValues (DSLContext context,
                                     Table<?> table,
                                     Field<String> row,
                                     Field<String> column,
-                                    Field<Double> value) {
+                                    Field<Double> value,
+                                    boolean isTemporary) {
     this.context = context;
     this.table = table;
     this.row = row;
     this.column = column;
     this.value = value;
+    this.isTemporary = isTemporary;
   }
 
   @Override
@@ -68,9 +71,11 @@ public class JooqBasedDataSourceValues extends AbstractDataSourceValues {
 
   @Override
   public void close () throws Exception {
-    if (log.isDebugEnabled ())
-      log.debug ("Dropping table " + table);
-    context.query ("DROP TABLE IF EXISTS {0}", table);
+    if(isTemporary){
+      if (log.isDebugEnabled ())
+        log.debug ("Dropping table " + table);
+      context.query ("DROP TABLE IF EXISTS {0}", table);
+    }
   }
 
   /* (non-Javadoc)
